+++ include/gestione_lavagna.h
@@ -14,16 +14,16 @@
 
 // Struttura per gli utenti
 struct Utente {
-    int32_t socket_utente;                  // File descriptor del socket
-    uint16_t porta_utente;                  // Numero di porta dell'utente
-    uint16_t attivo;                        // Un utente è attivo solo dopo l'HELLO
+    int socket_utente;                  // File descriptor del socket
+    int porta_utente;                  // Numero di porta dell'utente
+    int attivo;                        // Un utente è attivo solo dopo l'HELLO
     time_t tempo_invio_ping;                // Istante di invio del ping (server in attesa di pong)
     struct Utente * successivo;             // Puntatore per linked list
 };
 
 // Struttura per la Card
 struct Card {                               
-    uint16_t id;                            // Id della card
+    int id;                            // Id della card
     enum Colonne colonna;                   // Stato della card
     char testo[LUNGHEZZA_TESTO];            // Descrizione attività
     struct Utente * utente;                 // Utente che ha in carico la card
@@ -36,24 +36,24 @@ void show_lavagna();
 
 // Gestione delle card
 struct Card * get_card_todo(); // Restituisce la prima card in todo
-void sposta_card(uint16_t id, enum Colonne nuova_colonna, struct Utente * nuovo_utente);
+void sposta_card(int id, enum Colonne nuova_colonna, struct Utente * nuovo_utente);
 void assegna_card(struct Utente * utente);
-uint16_t get_prossimo_card_id_libero(); // Restituisce il primo id libero
+int get_prossimo_card_id_libero(); // Restituisce il primo id libero
 void crea_card(enum Colonne colonna, char * testo);
 
 // Gestione degli utenti
 void resetta_timer_lavoro_utente(struct Utente * utente);
 void interrompi_lavoro_utente(struct Utente * utente);
-void get_porte_attive(uint16_t * porte_utenti, uint16_t * num_utenti, struct Utente * utente_richiedente);
-uint16_t is_utente_occupato(struct Utente * utente); // 1 se occupato, 0 se libero
+void get_porte_attive(int * porte_utenti, int * num_utenti, struct Utente * utente_richiedente);
+int is_utente_occupato(struct Utente * utente); // 1 se occupato, 0 se libero
 struct Utente * get_utente_libero_min_porta();
 void invia_lista_utenti(struct Utente * utente);
-uint16_t conta_utenti_connessi(); // Restituisce il numero di utenti connessi
+int conta_utenti_connessi(); // Restituisce il numero di utenti connessi
 
 // Operazioni sugli utenti
 void disattiva_utente(struct Utente * utente);
-void attiva_utente(struct Utente * utente, uint16_t porta_utente);
+void attiva_utente(struct Utente * utente, int porta_utente);
 struct Utente * distruggi_utente(struct Utente * utente, struct Utente * precedente); // Restituisce il successivo nella linked list
-void crea_utente(int32_t socket_utente);
+void crea_utente(int socket_utente);
 
 #endif
