+++ include/protocollo_lavagna.h
@@ -13,7 +13,7 @@
 #include "gestione_lavagna.h"
 #include "common.h"
 
-void invia_messaggio(int32_t socket_fd, enum Comandi_lavagna_utente comando, struct Card * card, uint16_t * porte, uint16_t num_utenti);
-int32_t ricevi_messaggio(int32_t socket_fd, struct Messaggio_utente_lavagna * msg); // Restituisce 0 per connessione chiusa, -1 per errori
+void invia_messaggio(int socket_fd, enum Comandi_lavagna_utente comando, struct Card * card, int * porte, int num_utenti);
+int ricevi_messaggio(int socket_fd, struct Messaggio_utente_lavagna * msg); // Restituisce 0 per connessione chiusa, -1 per errori
 
 #endif
