+++ include/protocollo_utente.h
@@ -13,7 +13,7 @@
 #include "gestione_utente.h"
 #include "common.h"
 
-void invia_messaggio(int32_t socket_fd, enum Comandi_utente_lavagna comando, uint16_t porta_utente, uint16_t id_card, enum Colonne colonna, char * testo);
-int32_t ricevi_messaggio(int32_t socket_fd, struct Messaggio_lavagna_utente * msg); // Restituisce 0 per connessione chiusa, -1 per errori
+void invia_messaggio(int socket_fd, enum Comandi_utente_lavagna comando, int porta_utente, int id_card, enum Colonne colonna, char * testo);
+int ricevi_messaggio(int socket_fd, struct Messaggio_lavagna_utente * msg); // Restituisce 0 per connessione chiusa, -1 per errori
 
 #endif
